requirements.txt
flask==3.0.0
openai==1.3.5
httpx==0.28.1
python-dotenv==1.0.1
PyPDF2==3.0.1
python-docx==0.8.11
pillow==10.3.0
pytesseract==0.3.10